requirements.txt
openai
redis
requests
python-dotenv
fastapi
httpx
tqdm
